[Bronze II] Title: 숫자, Time: 264 ms, Memory: 22532 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/10093. 숫자/숫자.java
@@ -9,22 +9,22 @@ public class Main {
         BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
 
         StringTokenizer st = new StringTokenizer(br.readLine());
-        int n = Integer.parseInt(st.nextToken());
-        int m = Integer.parseInt(st.nextToken());
+        long n = Long.parseLong(st.nextToken());
+        long m = Long.parseLong(st.nextToken());
 
         if(n < m) {
             bw.write(m-n-1 + "\n");
 
-            for(int i = n +1; i<m; i++) {
+            for(long i = n + 1; i<m; i++) {
                 bw.write(i + " ");
             }
         } else if(n > m) {
             bw.write(n-m-1 + "\n");
 
-            for(int i = m +1; i<n; i++) {
+            for(long i = m +1; i<n; i++) {
                 bw.write(i + " ");
             }
-        } else if(n ==m) {
+        } else if(n == m) {
             bw.write("0");
         }
